feat: UserRepository 순환참조 문제 해결을 위한 인터페이스 기반 의존성 주입 및 UserService TS 마이그레이션
